refactor: merge data provider classes
