+++ demo/big.html
@@ -79,7 +79,7 @@
         ];
         LineUpJS.deriveColors(desc);
 
-        const p = new LineUpJS.LocalDataProvider(arr, desc, {
+        const p = new LineUpJS.DataProvider(arr, desc, {
           grouping: true,
           // taskExecutor: 'direct',
           taskExecutor: 'scheduled',
