+++ demo/big_d.html
@@ -67,7 +67,7 @@
           },
         ];
 
-        const p = new LineUpJS.LocalDataProvider(arr, desc, {
+        const p = new LineUpJS.DataProvider(arr, desc, {
           grouping: true,
         });
         p.deriveDefault();
