+++ demo/group.html
@@ -48,7 +48,7 @@
           },
         ];
 
-        const p = new LineUpJS.LocalDataProvider(arr, desc, { grouping: true });
+        const p = new LineUpJS.DataProvider(arr, desc, { grouping: true });
         p.deriveDefault();
         p.getLastRanking().insert(p.create({ type: 'selection', label: 'S' }), 1);
         p.getLastRanking().push(p.create(desc[1]));
