+++ src/provider/index.ts
@@ -1,5 +1,3 @@
-export { default as DataProvider } from './ADataProvider';
-export { default as LocalDataProvider } from './LocalDataProvider';
-export * from './LocalDataProvider';
+export { default as DataProvider } from './DataProvider';
 export { deriveColors, deriveColumnDescriptions, exportRanking, exportTable } from './utils';
 export * from './interfaces';
